Allow recursive calls for nested lams.
Right now it is not possible to have a `.fun a (n: .Nat)(...) = a n (..);` call it self recursively, as the partial evaluation for `n` will recurse endlessly.
This doesn't have to be this way.
Since we start rewriting the body by creating a new mutable, we can just cache this mutable immediately and use this for the recursive app of exactly the mutable + argument combination.